+++ Backend/Functionality/Organising.py
@@ -3,6 +3,7 @@ import logging
 from AiOrchestration.AiOrchestrator import AiOrchestrator
 from Data.CategoryManagement import CategoryManagement
 from Data.Configuration import Configuration
+from Data.Files.StorageMethodology import StorageMethodology
 from Data.NodeDatabaseManagement import NodeDatabaseManagement as nodeDB
 from Data.UserContextManagement import UserContextManagement
 from Utilities.Contexts import set_functionality_context
@@ -25,9 +26,23 @@ class Organising:
         return file_references
 
     @staticmethod
-    def store_prompt_data(user_prompt: str,
-                          response_message: str,
-                          category: str):
+    def save_file(content: str, category: str, file_path: str, overwrite=True):
+        """
+        Stores the system in the database, both the content in storage and the file node representation
+
+        :parameter content: As in file content
+        :parameter category: With respect to files the folder files are stored in
+        :parameter file_path: the full file address including folder directory and file name with extension.
+        """
+        nodeDB().create_file_node(category, file_path)
+        StorageMethodology().select().save_file(content, file_path, overwrite=overwrite)
+
+    @staticmethod
+    def store_prompt_data(
+        user_prompt: str,
+        response_message: str,
+        category: str
+    ):
         """
         ToDo: Look into celery and async processing